+++ src/main/java/com/nemonotfound/nemosmossyblocks/block/ModBlocks.java
@@ -30,6 +30,7 @@ public class ModBlocks {
     public static final Block MOSSY_ACACIA_FENCE = registerBlock("mossy_acacia_fence", new FenceBlock(AbstractBlock.Settings.copy(ACACIA_FENCE)));
     public static final Block MOSSY_ACACIA_FENCE_GATE = registerBlock("mossy_acacia_fence_gate", new FenceGateBlock(WoodType.ACACIA, AbstractBlock.Settings.copy(ACACIA_FENCE_GATE)));
     public static final Block MOSSY_ACACIA_DOOR = registerBlock("mossy_acacia_door", new DoorBlock(BlockSetType.ACACIA, AbstractBlock.Settings.copy(ACACIA_DOOR)));
+    public static final Block MOSSY_ACACIA_TRAPDOOR = registerBlock("mossy_acacia_trapdoor", new TrapdoorBlock(BlockSetType.ACACIA, AbstractBlock.Settings.copy(ACACIA_TRAPDOOR)));
     public static final Block MOSSY_ACACIA_PRESSURE_PLATE = registerBlock("mossy_acacia_pressure_plate",
             new PressurePlateBlock(BlockSetType.ACACIA, AbstractBlock.Settings.copy(ACACIA_PRESSURE_PLATE)));
     public static final Block MOSSY_ACACIA_BUTTON = registerBlock("mossy_acacia_button", Blocks.createWoodenButtonBlock(BlockSetType.ACACIA));
@@ -43,6 +44,7 @@ public class ModBlocks {
     public static final Block MOSSY_BAMBOO_FENCE = registerBlock("mossy_bamboo_fence", new FenceBlock(AbstractBlock.Settings.copy(BAMBOO_FENCE)));
     public static final Block MOSSY_BAMBOO_FENCE_GATE = registerBlock("mossy_bamboo_fence_gate", new FenceGateBlock(WoodType.BAMBOO, AbstractBlock.Settings.copy(BAMBOO_FENCE_GATE)));
     public static final Block MOSSY_BAMBOO_DOOR = registerBlock("mossy_bamboo_door", new DoorBlock(BlockSetType.BAMBOO, AbstractBlock.Settings.copy(BAMBOO_DOOR)));
+    public static final Block MOSSY_BAMBOO_TRAPDOOR = registerBlock("mossy_bamboo_trapdoor", new TrapdoorBlock(BlockSetType.BAMBOO, AbstractBlock.Settings.copy(BAMBOO_TRAPDOOR)));
     public static final Block MOSSY_BAMBOO_PRESSURE_PLATE = registerBlock("mossy_bamboo_pressure_plate",
             new PressurePlateBlock(BlockSetType.BAMBOO, AbstractBlock.Settings.copy(BAMBOO_PRESSURE_PLATE)));
     public static final Block MOSSY_BAMBOO_BUTTON = registerBlock("mossy_bamboo_button", Blocks.createWoodenButtonBlock(BlockSetType.BAMBOO));
@@ -54,6 +56,7 @@ public class ModBlocks {
     public static final Block MOSSY_BIRCH_FENCE = registerBlock("mossy_birch_fence", new FenceBlock(AbstractBlock.Settings.copy(BIRCH_FENCE)));
     public static final Block MOSSY_BIRCH_FENCE_GATE = registerBlock("mossy_birch_fence_gate", new FenceGateBlock(WoodType.BIRCH, AbstractBlock.Settings.copy(BIRCH_FENCE_GATE)));
     public static final Block MOSSY_BIRCH_DOOR = registerBlock("mossy_birch_door", new DoorBlock(BlockSetType.BIRCH, AbstractBlock.Settings.copy(BIRCH_DOOR)));
+    public static final Block MOSSY_BIRCH_TRAPDOOR = registerBlock("mossy_birch_trapdoor", new TrapdoorBlock(BlockSetType.BIRCH, AbstractBlock.Settings.copy(BIRCH_TRAPDOOR)));
     public static final Block MOSSY_BIRCH_PRESSURE_PLATE = registerBlock("mossy_birch_pressure_plate",
             new PressurePlateBlock(BlockSetType.BIRCH, AbstractBlock.Settings.copy(BIRCH_PRESSURE_PLATE)));
     public static final Block MOSSY_BIRCH_BUTTON = registerBlock("mossy_birch_button", Blocks.createWoodenButtonBlock(BlockSetType.BIRCH));
@@ -65,6 +68,7 @@ public class ModBlocks {
     public static final Block MOSSY_CHERRY_FENCE = registerBlock("mossy_cherry_fence", new FenceBlock(AbstractBlock.Settings.copy(CHERRY_FENCE)));
     public static final Block MOSSY_CHERRY_FENCE_GATE = registerBlock("mossy_cherry_fence_gate", new FenceGateBlock(WoodType.CHERRY, AbstractBlock.Settings.copy(CHERRY_FENCE_GATE)));
     public static final Block MOSSY_CHERRY_DOOR = registerBlock("mossy_cherry_door", new DoorBlock(BlockSetType.CHERRY, AbstractBlock.Settings.copy(CHERRY_DOOR)));
+    public static final Block MOSSY_CHERRY_TRAPDOOR = registerBlock("mossy_cherry_trapdoor", new TrapdoorBlock(BlockSetType.CHERRY, AbstractBlock.Settings.copy(CHERRY_TRAPDOOR)));
     public static final Block MOSSY_CHERRY_PRESSURE_PLATE = registerBlock("mossy_cherry_pressure_plate",
             new PressurePlateBlock(BlockSetType.CHERRY, AbstractBlock.Settings.copy(CHERRY_PRESSURE_PLATE)));
     public static final Block MOSSY_CHERRY_BUTTON = registerBlock("mossy_cherry_button", Blocks.createWoodenButtonBlock(BlockSetType.CHERRY));
@@ -76,6 +80,7 @@ public class ModBlocks {
     public static final Block MOSSY_CRIMSON_FENCE = registerBlock("mossy_crimson_fence", new FenceBlock(AbstractBlock.Settings.copy(CRIMSON_FENCE)));
     public static final Block MOSSY_CRIMSON_FENCE_GATE = registerBlock("mossy_crimson_fence_gate", new FenceGateBlock(WoodType.CRIMSON, AbstractBlock.Settings.copy(CRIMSON_FENCE_GATE)));
     public static final Block MOSSY_CRIMSON_DOOR = registerBlock("mossy_crimson_door", new DoorBlock(BlockSetType.CRIMSON, AbstractBlock.Settings.copy(CRIMSON_DOOR)));
+    public static final Block MOSSY_CRIMSON_TRAPDOOR = registerBlock("mossy_crimson_trapdoor", new TrapdoorBlock(BlockSetType.CRIMSON, AbstractBlock.Settings.copy(CRIMSON_TRAPDOOR)));
     public static final Block MOSSY_CRIMSON_PRESSURE_PLATE = registerBlock("mossy_crimson_pressure_plate",
             new PressurePlateBlock(BlockSetType.CRIMSON, AbstractBlock.Settings.copy(CRIMSON_PRESSURE_PLATE)));
     public static final Block MOSSY_CRIMSON_BUTTON = registerBlock("mossy_crimson_button", Blocks.createWoodenButtonBlock(BlockSetType.CRIMSON));
@@ -87,6 +92,7 @@ public class ModBlocks {
     public static final Block MOSSY_DARK_OAK_FENCE = registerBlock("mossy_dark_oak_fence", new FenceBlock(AbstractBlock.Settings.copy(DARK_OAK_FENCE)));
     public static final Block MOSSY_DARK_OAK_FENCE_GATE = registerBlock("mossy_dark_oak_fence_gate", new FenceGateBlock(WoodType.DARK_OAK, AbstractBlock.Settings.copy(DARK_OAK_FENCE_GATE)));
     public static final Block MOSSY_DARK_OAK_DOOR = registerBlock("mossy_dark_oak_door", new DoorBlock(BlockSetType.DARK_OAK, AbstractBlock.Settings.copy(DARK_OAK_DOOR)));
+    public static final Block MOSSY_DARK_OAK_TRAPDOOR = registerBlock("mossy_dark_oak_trapdoor", new TrapdoorBlock(BlockSetType.DARK_OAK, AbstractBlock.Settings.copy(DARK_OAK_TRAPDOOR)));
     public static final Block MOSSY_DARK_OAK_PRESSURE_PLATE = registerBlock("mossy_dark_oak_pressure_plate",
             new PressurePlateBlock(BlockSetType.DARK_OAK, AbstractBlock.Settings.copy(DARK_OAK_PRESSURE_PLATE)));
     public static final Block MOSSY_DARK_OAK_BUTTON = registerBlock("mossy_dark_oak_button", Blocks.createWoodenButtonBlock(BlockSetType.DARK_OAK));
@@ -98,6 +104,7 @@ public class ModBlocks {
     public static final Block MOSSY_JUNGLE_FENCE = registerBlock("mossy_jungle_fence", new FenceBlock(AbstractBlock.Settings.copy(JUNGLE_FENCE)));
     public static final Block MOSSY_JUNGLE_FENCE_GATE = registerBlock("mossy_jungle_fence_gate", new FenceGateBlock(WoodType.JUNGLE, AbstractBlock.Settings.copy(JUNGLE_FENCE_GATE)));
     public static final Block MOSSY_JUNGLE_DOOR = registerBlock("mossy_jungle_door", new DoorBlock(BlockSetType.JUNGLE, AbstractBlock.Settings.copy(JUNGLE_DOOR)));
+    public static final Block MOSSY_JUNGLE_TRAPDOOR = registerBlock("mossy_jungle_trapdoor", new TrapdoorBlock(BlockSetType.JUNGLE, AbstractBlock.Settings.copy(JUNGLE_TRAPDOOR)));
     public static final Block MOSSY_JUNGLE_PRESSURE_PLATE = registerBlock("mossy_jungle_pressure_plate",
             new PressurePlateBlock(BlockSetType.JUNGLE, AbstractBlock.Settings.copy(JUNGLE_PRESSURE_PLATE)));
     public static final Block MOSSY_JUNGLE_BUTTON = registerBlock("mossy_jungle_button",
@@ -110,6 +117,7 @@ public class ModBlocks {
     public static final Block MOSSY_MANGROVE_FENCE = registerBlock("mossy_mangrove_fence", new FenceBlock(AbstractBlock.Settings.copy(MANGROVE_FENCE)));
     public static final Block MOSSY_MANGROVE_FENCE_GATE = registerBlock("mossy_mangrove_fence_gate", new FenceGateBlock(WoodType.MANGROVE, AbstractBlock.Settings.copy(MANGROVE_FENCE_GATE)));
     public static final Block MOSSY_MANGROVE_DOOR = registerBlock("mossy_mangrove_door", new DoorBlock(BlockSetType.MANGROVE, AbstractBlock.Settings.copy(MANGROVE_DOOR)));
+    public static final Block MOSSY_MANGROVE_TRAPDOOR = registerBlock("mossy_mangrove_trapdoor", new TrapdoorBlock(BlockSetType.MANGROVE, AbstractBlock.Settings.copy(MANGROVE_TRAPDOOR)));
     public static final Block MOSSY_MANGROVE_PRESSURE_PLATE = registerBlock("mossy_mangrove_pressure_plate",
             new PressurePlateBlock(BlockSetType.MANGROVE, AbstractBlock.Settings.copy(MANGROVE_PRESSURE_PLATE)));
     public static final Block MOSSY_MANGROVE_BUTTON = registerBlock("mossy_mangrove_button", Blocks.createWoodenButtonBlock(BlockSetType.MANGROVE));
@@ -118,14 +126,13 @@ public class ModBlocks {
     public static final Block MOSSY_OAK_PLANKS = registerBlock("mossy_oak_planks", new Block(AbstractBlock.Settings.copy(OAK_PLANKS)));
     public static final Block MOSSY_OAK_STAIRS = registerBlock("mossy_oak_stairs", Blocks.createStairsBlock(MOSSY_OAK_PLANKS));
     public static final Block MOSSY_OAK_SLAB = registerBlock("mossy_oak_slab", new SlabBlock(MOSSY_OAK_PLANKS.getSettings()));
-    public static final Block MOSSY_OAK_FENCE = registerBlock("mossy_oak_fence",
-            new FenceBlock(AbstractBlock.Settings.copy(OAK_FENCE)));
-    public static final Block MOSSY_OAK_FENCE_GATE = registerBlock("mossy_oak_fence_gate",
-            new FenceGateBlock(WoodType.OAK, AbstractBlock.Settings.copy(OAK_FENCE_GATE)));
+    public static final Block MOSSY_OAK_FENCE = registerBlock("mossy_oak_fence", new FenceBlock(AbstractBlock.Settings.copy(OAK_FENCE)));
+    public static final Block MOSSY_OAK_FENCE_GATE = registerBlock("mossy_oak_fence_gate", new FenceGateBlock(WoodType.OAK, AbstractBlock.Settings.copy(OAK_FENCE_GATE)));
     public static final Block MOSSY_OAK_PRESSURE_PLATE = registerBlock("mossy_oak_pressure_plate",
             new PressurePlateBlock(BlockSetType.OAK, AbstractBlock.Settings.copy(OAK_PRESSURE_PLATE)));
     public static final Block MOSSY_OAK_BUTTON = registerBlock("mossy_oak_button", Blocks.createWoodenButtonBlock(BlockSetType.OAK));
     public static final Block MOSSY_OAK_DOOR = registerBlock("mossy_oak_door", new DoorBlock(BlockSetType.OAK, AbstractBlock.Settings.copy(OAK_DOOR)));
+    public static final Block MOSSY_OAK_TRAPDOOR = registerBlock("mossy_oak_trapdoor", new TrapdoorBlock(BlockSetType.OAK, AbstractBlock.Settings.copy(OAK_TRAPDOOR)));
     public static final Block MOSSY_SPRUCE_LOG = registerBlock("mossy_spruce_log", Blocks.createLogBlock(MapColor.SPRUCE_BROWN, MapColor.BROWN));
     public static final Block MOSSY_SPRUCE_WOOD = registerBlock("mossy_spruce_wood", new PillarBlock(AbstractBlock.Settings.copy(SPRUCE_WOOD)));
     public static final Block MOSSY_SPRUCE_PLANKS = registerBlock("mossy_spruce_planks", new Block(AbstractBlock.Settings.copy(SPRUCE_PLANKS)));
@@ -134,6 +141,7 @@ public class ModBlocks {
     public static final Block MOSSY_SPRUCE_FENCE = registerBlock("mossy_spruce_fence", new FenceBlock(AbstractBlock.Settings.copy(SPRUCE_FENCE)));
     public static final Block MOSSY_SPRUCE_FENCE_GATE = registerBlock("mossy_spruce_fence_gate", new FenceGateBlock(WoodType.SPRUCE, AbstractBlock.Settings.copy(SPRUCE_FENCE_GATE)));
     public static final Block MOSSY_SPRUCE_DOOR = registerBlock("mossy_spruce_door", new DoorBlock(BlockSetType.SPRUCE, AbstractBlock.Settings.copy(SPRUCE_DOOR)));
+    public static final Block MOSSY_SPRUCE_TRAPDOOR = registerBlock("mossy_spruce_trapdoor", new TrapdoorBlock(BlockSetType.SPRUCE, AbstractBlock.Settings.copy(SPRUCE_TRAPDOOR)));
     public static final Block MOSSY_SPRUCE_PRESSURE_PLATE = registerBlock("mossy_spruce_pressure_plate",
             new PressurePlateBlock(BlockSetType.SPRUCE, AbstractBlock.Settings.copy(SPRUCE_PRESSURE_PLATE)));
     public static final Block MOSSY_SPRUCE_BUTTON = registerBlock("mossy_spruce_button", Blocks.createWoodenButtonBlock(BlockSetType.SPRUCE));
@@ -145,6 +153,7 @@ public class ModBlocks {
     public static final Block MOSSY_WARPED_FENCE = registerBlock("mossy_warped_fence", new FenceBlock(AbstractBlock.Settings.copy(WARPED_FENCE)));
     public static final Block MOSSY_WARPED_FENCE_GATE = registerBlock("mossy_warped_fence_gate", new FenceGateBlock(WoodType.WARPED, AbstractBlock.Settings.copy(WARPED_FENCE_GATE)));
     public static final Block MOSSY_WARPED_DOOR = registerBlock("mossy_warped_door", new DoorBlock(BlockSetType.WARPED, AbstractBlock.Settings.copy(WARPED_DOOR)));
+    public static final Block MOSSY_WARPED_TRAPDOOR = registerBlock("mossy_warped_trapdoor", new TrapdoorBlock(BlockSetType.WARPED, AbstractBlock.Settings.copy(WARPED_TRAPDOOR)));
     public static final Block MOSSY_WARPED_PRESSURE_PLATE = registerBlock("mossy_warped_pressure_plate",
             new PressurePlateBlock(BlockSetType.WARPED, AbstractBlock.Settings.copy(WARPED_PRESSURE_PLATE)));
     public static final Block MOSSY_WARPED_BUTTON = registerBlock("mossy_warped_button", Blocks.createWoodenButtonBlock(BlockSetType.WARPED));
@@ -258,6 +267,7 @@ public class ModBlocks {
             ModItemGroups.NEMOS_MOSSY_BLOCKS);
     public static final Block MOSSY_IRON_BARS = registerBlock("mossy_iron_bars", new PaneBlock(IRON_BARS.getSettings()));
     public static final Block MOSSY_IRON_DOOR = registerBlock("mossy_iron_door", new DoorBlock(BlockSetType.IRON, AbstractBlock.Settings.copy(IRON_DOOR)));
+    public static final Block MOSSY_IRON_TRAPDOOR = registerBlock("mossy_iron_trapdoor", new TrapdoorBlock(BlockSetType.IRON, AbstractBlock.Settings.copy(IRON_TRAPDOOR)));
     public static final Block MOSSY_CHISELED_STONE_BRICKS = registerBlock("mossy_chiseled_stone_bricks",
             new Block(AbstractBlock.Settings.copy(CHISELED_STONE_BRICKS)));
     public static final Block INFESTED_MOSSY_CHISELED_STONE_BRICKS = registerBlock("infested_mossy_chiseled_stone_bricks",