+++ src/main/java/com/nemonotfound/nemosmossyblocks/datagen/ModelProvider.java
@@ -168,6 +168,19 @@ public class ModelProvider extends FabricModelProvider {
         blockStateModelGenerator.registerDoor(ModBlocks.MOSSY_CRIMSON_DOOR);
         blockStateModelGenerator.registerDoor(ModBlocks.MOSSY_WARPED_DOOR);
         blockStateModelGenerator.registerDoor(ModBlocks.MOSSY_IRON_DOOR);
+
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_OAK_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_SPRUCE_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_BIRCH_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_JUNGLE_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_ACACIA_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_DARK_OAK_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_MANGROVE_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_CHERRY_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_BAMBOO_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_CRIMSON_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_WARPED_TRAPDOOR);
+        blockStateModelGenerator.registerTrapdoor(ModBlocks.MOSSY_IRON_TRAPDOOR);
     }
 
     @Override