+++ src/main/java/com/nemonotfound/nemosmossyblocks/datagen/langdatagen/EnglishLanguageProvider.java
@@ -32,6 +32,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_OAK_FENCE, "Mossy Oak Fence");
         translationBuilder.add(ModItems.MOSSY_OAK_FENCE_GATE, "Mossy Oak Fence Gate");
         translationBuilder.add(ModItems.MOSSY_OAK_DOOR, "Mossy Oak Door");
+        translationBuilder.add(ModItems.MOSSY_OAK_TRAPDOOR, "Mossy Oak Trapdoor");
         translationBuilder.add(ModItems.MOSSY_OAK_PRESSURE_PLATE, "Mossy Oak Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_OAK_BUTTON, "Mossy Oak Button");
         translationBuilder.add(ModItems.MOSSY_SPRUCE_LOG, "Mossy Spruce Log");
@@ -42,6 +43,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_SPRUCE_FENCE, "Mossy Spruce Fence");
         translationBuilder.add(ModItems.MOSSY_SPRUCE_FENCE_GATE, "Mossy Spruce Fence Gate");
         translationBuilder.add(ModItems.MOSSY_SPRUCE_DOOR, "Mossy Spruce Door");
+        translationBuilder.add(ModItems.MOSSY_SPRUCE_TRAPDOOR, "Mossy Spruce Trapdoor");
         translationBuilder.add(ModItems.MOSSY_SPRUCE_PRESSURE_PLATE, "Mossy Spruce Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_SPRUCE_BUTTON, "Mossy Spruce Button");
         translationBuilder.add(ModItems.MOSSY_BIRCH_LOG, "Mossy Birch Log");
@@ -52,6 +54,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_BIRCH_FENCE, "Mossy Birch Fence");
         translationBuilder.add(ModItems.MOSSY_BIRCH_FENCE_GATE, "Mossy Birch Fence Gate");
         translationBuilder.add(ModItems.MOSSY_BIRCH_DOOR, "Mossy Birch Door");
+        translationBuilder.add(ModItems.MOSSY_BIRCH_TRAPDOOR, "Mossy Birch Trapdoor");
         translationBuilder.add(ModItems.MOSSY_BIRCH_PRESSURE_PLATE, "Mossy Birch Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_BIRCH_BUTTON, "Mossy Birch Button");
         translationBuilder.add(ModItems.MOSSY_JUNGLE_LOG, "Mossy Jungle Log");
@@ -62,6 +65,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_JUNGLE_FENCE, "Mossy Jungle Fence");
         translationBuilder.add(ModItems.MOSSY_JUNGLE_FENCE_GATE, "Mossy Jungle Fence Gate");
         translationBuilder.add(ModItems.MOSSY_JUNGLE_DOOR, "Mossy Jungle Door");
+        translationBuilder.add(ModItems.MOSSY_JUNGLE_TRAPDOOR, "Mossy Jungle Trapdoor");
         translationBuilder.add(ModItems.MOSSY_JUNGLE_PRESSURE_PLATE, "Mossy Jungle Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_JUNGLE_BUTTON, "Mossy Jungle Button");
         translationBuilder.add(ModItems.MOSSY_ACACIA_LOG, "Mossy Acacia Log");
@@ -72,6 +76,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_ACACIA_FENCE, "Mossy Acacia Fence");
         translationBuilder.add(ModItems.MOSSY_ACACIA_FENCE_GATE, "Mossy Acacia Fence Gate");
         translationBuilder.add(ModItems.MOSSY_ACACIA_DOOR, "Mossy Acacia Door");
+        translationBuilder.add(ModItems.MOSSY_ACACIA_TRAPDOOR, "Mossy Acacia Trapdoor");
         translationBuilder.add(ModItems.MOSSY_ACACIA_PRESSURE_PLATE, "Mossy Acacia Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_ACACIA_BUTTON, "Mossy Acacia Button");
         translationBuilder.add(ModItems.MOSSY_DARK_OAK_LOG, "Mossy Dark Oak Log");
@@ -82,6 +87,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_DARK_OAK_FENCE, "Mossy Dark Oak Fence");
         translationBuilder.add(ModItems.MOSSY_DARK_OAK_FENCE_GATE, "Mossy Dark Oak Fence Gate");
         translationBuilder.add(ModItems.MOSSY_DARK_OAK_DOOR, "Mossy Dark Oak Door");
+        translationBuilder.add(ModItems.MOSSY_DARK_OAK_TRAPDOOR, "Mossy Dark Oak Trapdoor");
         translationBuilder.add(ModItems.MOSSY_DARK_OAK_PRESSURE_PLATE, "Mossy Dark Oak Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_DARK_OAK_BUTTON, "Mossy Dark Oak Button");
         translationBuilder.add(ModItems.MOSSY_MANGROVE_LOG, "Mossy Mangrove Log");
@@ -92,6 +98,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_MANGROVE_FENCE, "Mossy Mangrove Fence");
         translationBuilder.add(ModItems.MOSSY_MANGROVE_FENCE_GATE, "Mossy Mangrove Fence Gate");
         translationBuilder.add(ModItems.MOSSY_MANGROVE_DOOR, "Mossy Mangrove Door");
+        translationBuilder.add(ModItems.MOSSY_MANGROVE_TRAPDOOR, "Mossy Mangrove Trapdoor");
         translationBuilder.add(ModItems.MOSSY_MANGROVE_PRESSURE_PLATE, "Mossy Mangrove Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_MANGROVE_BUTTON, "Mossy Mangrove Button");
         translationBuilder.add(ModItems.MOSSY_CHERRY_LOG, "Mossy Cherry Log");
@@ -102,6 +109,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_CHERRY_FENCE, "Mossy Cherry Fence");
         translationBuilder.add(ModItems.MOSSY_CHERRY_FENCE_GATE, "Mossy Cherry Fence Gate");
         translationBuilder.add(ModItems.MOSSY_CHERRY_DOOR, "Mossy Cherry Door");
+        translationBuilder.add(ModItems.MOSSY_CHERRY_TRAPDOOR, "Mossy Cherry Trapdoor");
         translationBuilder.add(ModItems.MOSSY_CHERRY_PRESSURE_PLATE, "Mossy Cherry Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_CHERRY_BUTTON, "Mossy Cherry Button");
         translationBuilder.add(ModItems.MOSSY_BAMBOO_BLOCK, "Mossy Bamboo Block");
@@ -114,6 +122,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_BAMBOO_FENCE, "Mossy Bamboo Fence");
         translationBuilder.add(ModItems.MOSSY_BAMBOO_FENCE_GATE, "Mossy Bamboo Fence Gate");
         translationBuilder.add(ModItems.MOSSY_BAMBOO_DOOR, "Mossy Bamboo Door");
+        translationBuilder.add(ModItems.MOSSY_BAMBOO_TRAPDOOR, "Mossy Bamboo Trapdoor");
         translationBuilder.add(ModItems.MOSSY_BAMBOO_PRESSURE_PLATE, "Mossy Bamboo Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_BAMBOO_BUTTON, "Mossy Bamboo Button");
         translationBuilder.add(ModItems.MOSSY_CRIMSON_STEM, "Mossy Crimson Stem");
@@ -124,6 +133,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_CRIMSON_FENCE, "Mossy Crimson Fence");
         translationBuilder.add(ModItems.MOSSY_CRIMSON_FENCE_GATE, "Mossy Crimson Fence Gate");
         translationBuilder.add(ModItems.MOSSY_CRIMSON_DOOR, "Mossy Crimson Door");
+        translationBuilder.add(ModItems.MOSSY_CRIMSON_TRAPDOOR, "Mossy Crimson Trapdoor");
         translationBuilder.add(ModItems.MOSSY_CRIMSON_PRESSURE_PLATE, "Mossy Crimson Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_CRIMSON_BUTTON, "Mossy Crimson Button");
         translationBuilder.add(ModItems.MOSSY_WARPED_STEM, "Mossy Warped Stem");
@@ -134,6 +144,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_WARPED_FENCE, "Mossy Warped Fence");
         translationBuilder.add(ModItems.MOSSY_WARPED_FENCE_GATE, "Mossy Warped Fence Gate");
         translationBuilder.add(ModItems.MOSSY_WARPED_DOOR, "Mossy Warped Door");
+        translationBuilder.add(ModItems.MOSSY_WARPED_TRAPDOOR, "Mossy Warped Trapdoor");
         translationBuilder.add(ModItems.MOSSY_WARPED_PRESSURE_PLATE, "Mossy Warped Pressure Plate");
         translationBuilder.add(ModItems.MOSSY_WARPED_BUTTON, "Mossy Warped Button");
         translationBuilder.add(ModItems.MOSSY_STONE, "Mossy Stone");
@@ -143,6 +154,7 @@ public class EnglishLanguageProvider extends FabricLanguageProvider {
         translationBuilder.add(ModItems.MOSSY_STONE_BUTTON, "Mossy Stone Button");
         translationBuilder.add(ModItems.MOSSY_IRON_BARS, "Mossy Iron Bars");
         translationBuilder.add(ModItems.MOSSY_IRON_DOOR, "Mossy Iron Door");
+        translationBuilder.add(ModItems.MOSSY_IRON_TRAPDOOR, "Mossy Iron Trapdoor");
 
         try {
             Optional<Path> optionalPath = dataOutput.getModContainer().findPath("en_us.json");