+++ src/main/java/com/nemonotfound/nemosmossyblocks/item/ModItemGroups.java
@@ -49,6 +49,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_OAK_FENCE);
                     entries.add(ModItems.MOSSY_OAK_FENCE_GATE);
                     entries.add(ModItems.MOSSY_OAK_DOOR);
+                    entries.add(ModItems.MOSSY_OAK_TRAPDOOR);
                     entries.add(ModItems.MOSSY_OAK_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_OAK_BUTTON);
                     entries.add(ModItems.MOSSY_SPRUCE_LOG);
@@ -59,6 +60,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_SPRUCE_FENCE);
                     entries.add(ModItems.MOSSY_SPRUCE_FENCE_GATE);
                     entries.add(ModItems.MOSSY_SPRUCE_DOOR);
+                    entries.add(ModItems.MOSSY_SPRUCE_TRAPDOOR);
                     entries.add(ModItems.MOSSY_SPRUCE_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_SPRUCE_BUTTON);
                     entries.add(ModItems.MOSSY_BIRCH_LOG);
@@ -69,6 +71,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_BIRCH_FENCE);
                     entries.add(ModItems.MOSSY_BIRCH_FENCE_GATE);
                     entries.add(ModItems.MOSSY_BIRCH_DOOR);
+                    entries.add(ModItems.MOSSY_BIRCH_TRAPDOOR);
                     entries.add(ModItems.MOSSY_BIRCH_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_BIRCH_BUTTON);
                     entries.add(ModItems.MOSSY_JUNGLE_LOG);
@@ -79,6 +82,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_JUNGLE_FENCE);
                     entries.add(ModItems.MOSSY_JUNGLE_FENCE_GATE);
                     entries.add(ModItems.MOSSY_JUNGLE_DOOR);
+                    entries.add(ModItems.MOSSY_JUNGLE_TRAPDOOR);
                     entries.add(ModItems.MOSSY_JUNGLE_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_JUNGLE_BUTTON);
                     entries.add(ModItems.MOSSY_ACACIA_LOG);
@@ -89,6 +93,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_ACACIA_FENCE);
                     entries.add(ModItems.MOSSY_ACACIA_FENCE_GATE);
                     entries.add(ModItems.MOSSY_ACACIA_DOOR);
+                    entries.add(ModItems.MOSSY_ACACIA_TRAPDOOR);
                     entries.add(ModItems.MOSSY_ACACIA_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_ACACIA_BUTTON);
                     entries.add(ModItems.MOSSY_DARK_OAK_LOG);
@@ -99,6 +104,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_DARK_OAK_FENCE);
                     entries.add(ModItems.MOSSY_DARK_OAK_FENCE_GATE);
                     entries.add(ModItems.MOSSY_DARK_OAK_DOOR);
+                    entries.add(ModItems.MOSSY_DARK_OAK_TRAPDOOR);
                     entries.add(ModItems.MOSSY_DARK_OAK_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_DARK_OAK_BUTTON);
                     entries.add(ModItems.MOSSY_MANGROVE_LOG);
@@ -109,6 +115,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_MANGROVE_FENCE);
                     entries.add(ModItems.MOSSY_MANGROVE_FENCE_GATE);
                     entries.add(ModItems.MOSSY_MANGROVE_DOOR);
+                    entries.add(ModItems.MOSSY_MANGROVE_TRAPDOOR);
                     entries.add(ModItems.MOSSY_MANGROVE_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_MANGROVE_BUTTON);
                     entries.add(ModItems.MOSSY_CHERRY_LOG);
@@ -120,6 +127,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_CHERRY_FENCE_GATE);
                     entries.add(ModItems.MOSSY_CHERRY_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_CHERRY_DOOR);
+                    entries.add(ModItems.MOSSY_CHERRY_TRAPDOOR);
                     entries.add(ModItems.MOSSY_CHERRY_BUTTON);
                     entries.add(ModItems.MOSSY_BAMBOO_BLOCK);
                     entries.add(ModItems.MOSSY_BAMBOO_PLANKS);
@@ -131,6 +139,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_BAMBOO_FENCE);
                     entries.add(ModItems.MOSSY_BAMBOO_FENCE_GATE);
                     entries.add(ModItems.MOSSY_BAMBOO_DOOR);
+                    entries.add(ModItems.MOSSY_BAMBOO_TRAPDOOR);
                     entries.add(ModItems.MOSSY_BAMBOO_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_BAMBOO_BUTTON);
                     entries.add(ModItems.MOSSY_CRIMSON_STEM);
@@ -141,6 +150,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_CRIMSON_FENCE);
                     entries.add(ModItems.MOSSY_CRIMSON_FENCE_GATE);
                     entries.add(ModItems.MOSSY_CRIMSON_DOOR);
+                    entries.add(ModItems.MOSSY_CRIMSON_TRAPDOOR);
                     entries.add(ModItems.MOSSY_CRIMSON_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_CRIMSON_BUTTON);
                     entries.add(ModItems.MOSSY_WARPED_STEM);
@@ -151,6 +161,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_WARPED_FENCE);
                     entries.add(ModItems.MOSSY_WARPED_FENCE_GATE);
                     entries.add(ModItems.MOSSY_WARPED_DOOR);
+                    entries.add(ModItems.MOSSY_WARPED_TRAPDOOR);
                     entries.add(ModItems.MOSSY_WARPED_PRESSURE_PLATE);
                     entries.add(ModItems.MOSSY_WARPED_BUTTON);
                     entries.add(ModItems.MOSSY_STONE);
@@ -166,6 +177,7 @@ public class ModItemGroups {
                     entries.add(ModItems.MOSSY_BRICK_WALL);
                     entries.add(ModItems.MOSSY_IRON_BARS);
                     entries.add(ModItems.MOSSY_IRON_DOOR);
+                    entries.add(ModItems.MOSSY_IRON_TRAPDOOR);
                 })
                 .build();
 
@@ -191,6 +203,7 @@ public class ModItemGroups {
         ItemGroupEvents.modifyEntriesEvent(ItemGroups.BUILDING_BLOCKS).register(entries -> {
             entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_BUTTON);
             entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_PRESSURE_PLATE);
+            entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_TRAPDOOR);
             entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_DOOR);
             entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_FENCE_GATE);
             entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_FENCE);
@@ -201,6 +214,7 @@ public class ModItemGroups {
             entries.addAfter(Items.OAK_BUTTON, ModItems.MOSSY_OAK_LOG);
             entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_BUTTON);
             entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_PRESSURE_PLATE);
+            entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_TRAPDOOR);
             entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_DOOR);
             entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_FENCE_GATE);
             entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_FENCE);
@@ -211,6 +225,7 @@ public class ModItemGroups {
             entries.addAfter(Items.SPRUCE_BUTTON, ModItems.MOSSY_SPRUCE_LOG);
             entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_BUTTON);
             entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_PRESSURE_PLATE);
+            entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_TRAPDOOR);
             entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_DOOR);
             entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_FENCE_GATE);
             entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_FENCE);
@@ -221,6 +236,7 @@ public class ModItemGroups {
             entries.addAfter(Items.BIRCH_BUTTON, ModItems.MOSSY_BIRCH_LOG);
             entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_BUTTON);
             entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_PRESSURE_PLATE);
+            entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_TRAPDOOR);
             entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_DOOR);
             entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_FENCE_GATE);
             entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_FENCE);
@@ -231,6 +247,7 @@ public class ModItemGroups {
             entries.addAfter(Items.JUNGLE_BUTTON, ModItems.MOSSY_JUNGLE_LOG);
             entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_BUTTON);
             entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_PRESSURE_PLATE);
+            entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_TRAPDOOR);
             entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_DOOR);
             entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_FENCE_GATE);
             entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_FENCE);
@@ -241,6 +258,7 @@ public class ModItemGroups {
             entries.addAfter(Items.ACACIA_BUTTON, ModItems.MOSSY_ACACIA_LOG);
             entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_BUTTON);
             entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_PRESSURE_PLATE);
+            entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_TRAPDOOR);
             entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_DOOR);
             entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_FENCE_GATE);
             entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_FENCE);
@@ -251,6 +269,7 @@ public class ModItemGroups {
             entries.addAfter(Items.DARK_OAK_BUTTON, ModItems.MOSSY_DARK_OAK_LOG);
             entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_BUTTON);
             entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_PRESSURE_PLATE);
+            entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_TRAPDOOR);
             entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_DOOR);
             entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_FENCE_GATE);
             entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_FENCE);
@@ -261,6 +280,7 @@ public class ModItemGroups {
             entries.addAfter(Items.MANGROVE_BUTTON, ModItems.MOSSY_MANGROVE_LOG);
             entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_BUTTON);
             entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_PRESSURE_PLATE);
+            entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_TRAPDOOR);
             entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_DOOR);
             entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_FENCE_GATE);
             entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_FENCE);
@@ -271,6 +291,7 @@ public class ModItemGroups {
             entries.addAfter(Items.CHERRY_BUTTON, ModItems.MOSSY_CHERRY_LOG);
             entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_BUTTON);
             entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_PRESSURE_PLATE);
+            entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_TRAPDOOR);
             entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_DOOR);
             entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_FENCE_GATE);
             entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_FENCE);
@@ -283,6 +304,7 @@ public class ModItemGroups {
             entries.addAfter(Items.BAMBOO_BUTTON, ModItems.MOSSY_BAMBOO_BLOCK);
             entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_BUTTON);
             entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_PRESSURE_PLATE);
+            entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_TRAPDOOR);
             entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_DOOR);
             entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_FENCE_GATE);
             entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_FENCE);
@@ -293,6 +315,7 @@ public class ModItemGroups {
             entries.addAfter(Items.CRIMSON_BUTTON, ModItems.MOSSY_CRIMSON_STEM);
             entries.addAfter(Items.WARPED_BUTTON, ModItems.MOSSY_WARPED_BUTTON);
             entries.addAfter(Items.WARPED_BUTTON, ModItems.MOSSY_WARPED_PRESSURE_PLATE);
+            entries.addAfter(Items.WARPED_BUTTON, ModItems.MOSSY_WARPED_TRAPDOOR);
             entries.addAfter(Items.WARPED_BUTTON, ModItems.MOSSY_WARPED_DOOR);
             entries.addAfter(Items.WARPED_BUTTON, ModItems.MOSSY_WARPED_FENCE_GATE);
             entries.addAfter(Items.WARPED_BUTTON, ModItems.MOSSY_WARPED_FENCE);
@@ -311,8 +334,9 @@ public class ModItemGroups {
             entries.addAfter(Items.BRICK_WALL, ModItems.MOSSY_BRICK_SLAB);
             entries.addAfter(Items.BRICK_WALL, ModItems.MOSSY_BRICK_STAIRS);
             entries.addAfter(Items.BRICK_WALL, ModItems.MOSSY_BRICKS);
-            entries.addAfter(Items.IRON_BARS, ModItems.MOSSY_IRON_BARS);
-            entries.addAfter(Items.IRON_DOOR, ModItems.MOSSY_IRON_DOOR);
+            entries.addAfter(Items.IRON_TRAPDOOR, ModItems.MOSSY_IRON_TRAPDOOR);
+            entries.addAfter(Items.IRON_TRAPDOOR, ModItems.MOSSY_IRON_DOOR);
+            entries.addAfter(Items.IRON_TRAPDOOR, ModItems.MOSSY_IRON_BARS);
         });
     }